Create QuizGame.java
The program will ask the user the first question.  The program will take the answer the user provides and check to see if they are right.
If they are right the program will generate another question. If they are wrong, it will give them one more try at that question then move on to the next question.
The program will ask a total of three questions.
After the three questions are asked, the program will display the results of number right and number wrong.